+add lib/Color/Color.dart +add lib/Components/Cardlist.dart +add widget_box.dart parameter -remove tempStatusbar(), tempStatusbar() in recentscreen.dart
