+++ flutter_application_1/lib/Color/Color.dart
@@ -1,0 +1,8 @@
+// ignore_for_file: constant_identifier_names
+
+import 'package:flutter/material.dart';
+
+
+const CHART_MINUS = Color.fromRGBO(0, 57, 164, 0.9);
+const CHART_PLUS = Color.fromRGBO(238, 71, 81, 0.9);
+
